Move client grant resource to client pkg
